+++ group-coordinator/src/main/java/org/apache/kafka/coordinator/group/api/assignor/GroupAssignment.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.coordinator.group.assignor;
+package org.apache.kafka.coordinator.group.api.assignor;
 
 import java.util.Map;
 import java.util.Objects;
